tools/requirements.txt
PyYAML==6.0.2
frigate==0.7.0
